created a controller for education form
